requirements.txt
flask
numpy
scipy
librosa
soundfile
pydub
openai-whisper
praat-parselmouth
torch
torchaudio
